Move `context.ts`
