+++ packages/seroval/src/tree/primitives.ts
@@ -1,7 +1,7 @@
 import assert from '../assert';
 import { Feature } from '../compat';
-import type { ParserContext } from '../context';
-import { createIndexedValue } from '../context';
+import type { ParserContext } from './context';
+import { createIndexedValue } from './context';
 import { serializeString } from '../string';
 import type { BigIntTypedArrayValue, TypedArrayValue } from '../types';
 import UnsupportedTypeError from './UnsupportedTypeError';
