+++ packages/seroval/src/tree/web-api.ts
@@ -1,6 +1,6 @@
 import assert from '../assert';
 import { Feature } from '../compat';
-import type { ParserContext } from '../context';
+import type { ParserContext } from './context';
 import { serializeString } from '../string';
 import UnsupportedTypeError from './UnsupportedTypeError';
 import { serializeArrayBuffer } from './primitives';
